chore: rename file to reflect its purpose
