+++ src/index.ts
@@ -2,13 +2,13 @@ import 'dotenv/config';
 import express from 'express';
 import compression from 'compression';
 import cors from 'cors';
-import rpc from './rpc';
+import api from './api';
 
 const app = express();
 const PORT = process.env.PORT || 3000;
 
 app.use(cors({ maxAge: 86400 }));
 app.use(compression());
-app.use('/', rpc);
+app.use('/', api);
 
 app.listen(PORT, () => console.log(`Listening at http://localhost:${PORT}`));
